refactor: extract exportToExcel method to a new file
